+++ core/src/main/java/de/bitbrain/braingdx/input/InputManager.java
@@ -1,0 +1,14 @@
+package de.bitbrain.braingdx.input;
+
+import com.badlogic.gdx.InputProcessor;
+import com.badlogic.gdx.controllers.ControllerListener;
+import de.bitbrain.braingdx.behavior.movement.Movement;
+import de.bitbrain.braingdx.behavior.movement.Orientation;
+
+/**
+ * Registers input events
+ */
+public interface InputManager {
+   void register(ControllerListener controllerListener);
+   void register(InputProcessor inputAdapter);
+}